request user login mechanism DRY
